requirements.txt
Django==1.9.6
github3.py==1.0.0a4
pycrypto==2.6.1
requests==2.6.0
six==1.10.0
uritemplate.py==0.3.0
wechat-sdk==0.6.3
xmltodict==0.9.2
